requirements.txt
cmake
dlib
face_recognition
numpy
opencv-python

astroid==2.4.2
click==7.1.2
cmake==3.18.2.post1
colorama==0.4.3
dlib==19.21.0
face-recognition==1.3.0
face-recognition-models==0.3.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.18.5
opencv-python==4.4.0.42
pillow==7.2.0
pylint==2.6.0
six==1.15.0
toml==0.10.1
wrapt==1.12.1